+++ R/cartography_chamber.R
@@ -13,8 +13,7 @@
 #' @param dists A distance matrix for the data frame.
 #' @param filtered_data The result of a function applied to the data frame; there should be one filter value per observation in the original data frame.
 #' @param cover_element_tests A list of membership test functions for a set of cover elements. In other words, each element of `cover_element_tests` is a function that returns `TRUE` or `FALSE` when given a filter value.
-#' @param method A string to pass to [hclust] to determine clustering method.
-#' @param global_clustering Whether you want clustering to happen in a global (all level visible) or local (only current level set visible) context
+#' @param clusterer A function which accepts a list of distance matrices as input, and returns the results of clustering done on each distance matrix. Defaults to `NULL`, meaning no all data in each bin will be lumped into a single cluster.
 #'
 #' @return A list of two dataframes, one with node data and one with edge data.
 #' @export
@@ -38,15 +37,13 @@
 #'   data = data,
 #'   dists = dist(data),
 #'   filtered_data = projx,
-#'   cover_element_tests = xcovercheck,
-#'   method = "single"
+#'   cover_element_tests = xcovercheck
 #' )
 create_mapper_object <- function(data,
                                  dists,
                                  filtered_data,
                                  cover_element_tests,
-                                 method = "none",
-                                 global_clustering = TRUE) {
+                                 clusterer = NULL) {
   if (!is.data.frame(data)) {
     stop("input data needs to be a data frame.")
   } else if (!all(sapply(cover_element_tests, typeof) == "closure")) {
@@ -75,11 +72,11 @@ create_mapper_object <- function(data,
 
   bins = create_bins(data, filtered_data, cover_element_tests)
 
-  if (method == "none") {
-    return(run_mapper(convert_to_clusters(bins), dists, binning = FALSE))
+  if (is.null(clusterer)) {
+    return(assemble_mapper_object(convert_to_clusters(bins), dists, binning = FALSE))
   } else {
-    clusters = get_clusters(bins, dists, method)
-    return(run_mapper(clusters, dists, binning = TRUE))
+    clusters = get_clusters(bins, dists, clusterer)
+    return(assemble_mapper_object(clusters, dists, binning = TRUE))
   }
 }
 
@@ -131,7 +128,7 @@ create_bins <- function(data, filtered_data, cover_element_tests) {
 #' @return A list of two dataframes, one with node data containing bin membership,
 #'  datapoints per cluster, and cluster dispersion, and one with edge data
 #'  containing sources, targets, and weights representing overlap strength.
-run_mapper <- function(binclust_data, dists, binning = TRUE) {
+assemble_mapper_object <- function(binclust_data, dists, binning = TRUE) {
 
   # basic node info
   num_vertices = max(binclust_data[[length(binclust_data)]])
@@ -157,7 +154,8 @@ run_mapper <- function(binclust_data, dists, binning = TRUE) {
       source = sources,
       target = targets,
       weight = edge_weights,
-      overlap = data_in_overlap
+      overlap_data = data_in_overlap,
+      overlap_size = sapply(overlaps, length)
     )
   } else {
     edges = data.frame(source = "", target = "")